load javascript and css via adjunct to be csp compliant (#24)
the inline script code is moved to a dedicated js file that is loaded
via st:adjunct
Inline javascript is a problem when one wants to apply content security
policies that have `script-src: 'self'`
Remove the page decorater that lead to loading the jsoneditor_min.js
file (500kb) on every page of Jenkins which wastes bandwidth and slows
down page loading. It is also a problem for the above mentioned CSP
setting.
The js file is now stored in the plugin itself and loaded by Jenkins
only when a json editor parameter is present when triggering a job or an
input step makes use of it.

Define some css styles in a file (loaded via adjunct) that overwrites
things from Jenkins so the rendered form looks nice and menus work
properly. Ensures that menus work with dark theme